# OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RevFIT.API",
    "version": "1.0"
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIRegisterViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/APIRegisterViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/APIRegisterViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIUserLoginModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/APIUserLoginModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/APIUserLoginModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringServiceResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringServiceResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringServiceResponseModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/workout/todays-workout": {
      "get": {
        "tags": [
          "Workout"
        ],
        "parameters": [
          {
            "name": "progrogramID",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "dateTime",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WodListServiceResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WodListServiceResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WodListServiceResponseModel"
                }
              }
            }
          }
        }
      }
    },
    "/workout/add-workout": {
      "post": {
        "tags": [
          "Workout"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutAddRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutAddRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WorkoutAddRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponseModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponseModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32ServiceResponseModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "APIRegisterViewModel": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "APIUserLoginModel": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ExcerciseProgram": {
        "type": "object",
        "properties": {
          "programId": {
            "type": "integer",
            "format": "int32"
          },
          "programName": {
            "type": "string",
            "nullable": true
          },
          "programTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "programDurationWeeks": {
            "type": "integer",
            "format": "int32"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "programType": {
            "$ref": "#/components/schemas/ProgramType"
          },
          "wods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wod"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Int32ServiceResponseModel": {
        "type": "object",
        "properties": {
          "data": {
            "type": "integer",
            "format": "int32"
          },
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
      "ProgramType": {
        "type": "object",
        "properties": {
          "programTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "excercisePrograms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExcerciseProgram"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScoreCalculationType": {
        "type": "object",
        "properties": {
          "scoreCalculationTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "workoutScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutScore"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScoreMeasureType": {
        "type": "object",
        "properties": {
          "measureTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "workoutScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutScore"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ScoreOrderType": {
        "type": "object",
        "properties": {
          "scoreOrderTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "workoutScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutScore"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StringServiceResponseModel": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string",
            "nullable": true
          },
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Wod": {
        "type": "object",
        "properties": {
          "wodId": {
            "type": "integer",
            "format": "int32"
          },
          "workoutName": {
            "type": "string",
            "nullable": true
          },
          "workoutDefinition": {
            "type": "string",
            "nullable": true
          },
          "programId": {
            "type": "integer",
            "format": "int32"
          },
          "scoreType": {
            "type": "string",
            "nullable": true
          },
          "scheduleDate": {
            "type": "string",
            "format": "date-time"
          },
          "isLive": {
            "type": "boolean"
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "program": {
            "$ref": "#/components/schemas/ExcerciseProgram"
          },
          "workoutScores": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkoutScore"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WodListServiceResponseModel": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Wod"
            },
            "nullable": true
          },
          "isSuccess": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WorkoutAddRequestModel": {
        "type": "object",
        "properties": {
          "wodId": {
            "type": "integer",
            "format": "int32"
          },
          "workoutName": {
            "type": "string",
            "nullable": true
          },
          "workoutDefinition": {
            "type": "string",
            "nullable": true
          },
          "programId": {
            "type": "integer",
            "format": "int32"
          },
          "measureTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "calcTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "orderTypeID": {
            "type": "integer",
            "format": "int32"
          },
          "set": {
            "type": "integer",
            "format": "int32"
          },
          "scoreType": {
            "type": "string",
            "nullable": true
          },
          "scheduleDate": {
            "type": "string",
            "format": "date-time"
          },
          "isLive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "WorkoutScore": {
        "type": "object",
        "properties": {
          "scoreId": {
            "type": "integer",
            "format": "int32"
          },
          "sets": {
            "type": "integer",
            "format": "int32"
          },
          "measureTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "scoreCalculationTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "scoreOrderId": {
            "type": "integer",
            "format": "int32"
          },
          "workoutId": {
            "type": "integer",
            "format": "int32"
          },
          "dateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "measureType": {
            "$ref": "#/components/schemas/ScoreMeasureType"
          },
          "scoreCalculationType": {
            "$ref": "#/components/schemas/ScoreCalculationType"
          },
          "scoreOrder": {
            "$ref": "#/components/schemas/ScoreOrderType"
          },
          "workout": {
            "$ref": "#/components/schemas/Wod"
          }
        },
        "additionalProperties": false
      }
    }
  }
}